\name{compute_hellinger_gammas}
\alias{compute_hellinger_gammas}
\title{Compute Gamma HD analitically}

\usage{compute_hellinger_gammas(hyperpar1, hyperpar2)}
\arguments{
  \item{hyperpar1}{Should be a named numeric array with 'shape' and 'rate' elements of 
the first Gamma distribution.}
  \item{hyperpar2}{Should be a named numeric array with 'shape' and 'rate' elements of 
the second Gamma distribution.}
}



\author{Thiago Guerrera Martins}





