\name{compute_hellinger_gaussians}
\alias{compute_hellinger_gaussians}
\title{Compute Gaussian HD analitically}

\usage{compute_hellinger_gaussians(hyperpar1, hyperpar2)}
\arguments{
  \item{hyperpar1}{Should be a named numeric array with 'mean' and 'prec' elements of 
the first Gaussian distribution.}
  \item{hyperpar2}{Should be a named numeric array with 'mean' and 'prec' elements of 
the second Gaussian distribution.}
}



\author{Thiago Guerrera Martins}





