\name{priorSens-package}
\alias{priorSens-package}
\alias{priorSens}
\docType{package}
\title{Prior sensitivity with respect to hyperparameters}
\description{Prior sensitivity with respect to hyperparameters in Latent Gaussian Models}
\details{
\tabular{ll}{Package: \tab priorSens\cr
Type: \tab Package\cr
Title: \tab Prior sensitivity with respect to hyperparameters\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-10-29\cr
Author: \tab Thiago Guerrera Martins\cr
Maintainer: \tab Thiago Guerrera Martins <thiago.guerrera@math.ntnu.no>\cr
License: \tab GPL (>= 2)\cr
Depends: \tab INLA, rootSolve, multicore, compiler, Rcpp\cr
LinkingTo: \tab Rcpp\cr
LazyLoad: \tab yes\cr}
}
\author{Thiago Guerrera Martins}

\keyword{ package }


